requirements.txt
absl-py==1.4.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astunparse==1.6.3
attrs==23.1.0
autograd==1.6.2
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cached-property==1.5.2
cachetools==5.3.1
cffi==1.15.1
charset-normalizer==3.2.0
cma==2.7.0
colorama==0.4.6
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.1.2
fastjsonschema==2.18.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
fsspec==2023.1.0
future==0.18.3
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.8.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jstyleson==0.0.2
jupyter-events==0.6.3
jupyter-server==1.24.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter_server_fileid==0.9.0
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.23.0
keras==2.11.0
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.5.2
matplotlib-inline==0.1.6
mistune==3.0.1
natsort==8.4.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.6.3
ninja==1.10.2.4
nncf==2.5.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.21.6
oauthlib==3.2.2
onnx==1.12.0
opencv-python==4.8.0.74
openvino==2023.0.0
openvino-dev==2023.0.0
openvino-telemetry==2022.3.0
opt-einsum==3.3.0
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
Pygments==2.15.1
pymoo==0.5.0
pyparsing==2.4.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.3.0
pywin32==306
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.2
singledispatchmethod==1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
texttable==1.6.7
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==4.0.8
wincertstore==0.2
wrapt==1.15.0
y-py==0.6.0
ypy-websocket==0.8.4
zipp==3.15.0
